Terminal - moved UI related things to its own class
